split screenmenu into 2 files (for 2 classes..). Few tricks for display for pi.
