+++ Makefile
@@ -13,7 +13,8 @@ OBJS	= src/main.o\
 		src/emulation.o\
 		src/KeyboardPi.o\
 		src/log.o\
-		src/SoundPi.o\
+		src/MainMenuWindows.o\
+		src/MainMenuWindowsSoundPi.o\
 		src/ScreenMenu.o\
 		src/Window.o\
 		res/SugarboxLogo.o\
